refactor: minimize diff and follow reth patterns
- Move base fee test to existing test module instead of separate file
- Remove unnecessary trait bound from EthChainSpec::Header
- Maintain backward compatibility by keeping original trait signature

This reduces the diff size and follows reth's testing conventions.
